backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9; sys_platform != "win32"
python-dotenv==1.0.1
anthropic==0.28.0
apscheduler==3.10.4
gunicorn==22.0.0
sqlalchemy==2.0.31
feedparser==6.0.11
requests==2.32.3
